feat: integrate OpenRouter Meta Llama 3.3, voice audio sync, decimal math support, and UI declutter

+++ backend/src/contracts/gameplay_engine.py
@@ -138,7 +138,7 @@ class HintResponse(BaseModel):
 
 
 class AnswerChallengeRequest(BaseModel):
-    answer: int = Field(ge=0, le=100_000)
+    answer: float = Field(ge=0, le=100_000)
 
 
 class AnswerChallengeResponse(BaseModel):
